requirements.txt
python-dotenv
requests
numpy
pandas
pulp
pytz
nordpool
ics